refactor: util 내 패키지 유형별로 분리
