+++ src/main/java/controller/UserController.java
@@ -10,8 +10,8 @@ import db.UserDatabase;
 import dto.LoginDto;
 import dto.UserCreateDto;
 import model.User;
-import util.RequestParser;
-import util.SessionManager;
+import util.web.RequestParser;
+import util.session.SessionManager;
 import webserver.HttpRequest;
 import webserver.HttpResponse;
