+++ src/main/java/util/session/SessionInfo.java
@@ -1,4 +1,4 @@
-package util;
+package util.session;
 
 import model.User;
