+++ src/main/java/util/session/SessionManager.java
@@ -1,7 +1,8 @@
-package util;
+package util.session;
 
 import constant.HttpHeader;
 import model.User;
+import util.web.RequestParser;
 import webserver.HttpRequest;
 
 import java.util.Map;
