+++ src/main/java/util/web/RequestMapper.java
@@ -1,4 +1,4 @@
-package util;
+package util.web;
 
 import annotation.RequestBody;
 import annotation.RequestMapping;
@@ -7,6 +7,7 @@ import constant.HttpStatus;
 import constant.ParamType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import util.ClassScanner;
 import webserver.HttpRequest;
 import webserver.HttpResponse;
