+++ src/main/java/webserver/HttpRequest.java
@@ -1,8 +1,7 @@
 package webserver;
 
 import constant.HttpHeader;
-import util.ObjectMapper;
-import util.RequestParser;
+import util.web.RequestParser;
 
 import java.io.BufferedReader;
 import java.io.IOException;
