+++ src/main/java/webserver/RequestHandler.java
@@ -6,7 +6,11 @@ import constant.MimeType;
 import exception.ResourceNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import util.*;
+import util.html.HtmlBuilder;
+import util.web.SharedData;
+import util.session.SessionManager;
+import util.web.RequestMapper;
+import util.web.ResourceLoader;
 
 import java.io.*;
 import java.lang.reflect.Method;
